% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_boc_all_series_last_date}
\alias{get_boc_all_series_last_date}
\title{Get series to series group link and date of last available data point}
\usage{
get_boc_all_series_last_date(quiet = TRUE)
}
\arguments{
\item{quiet}{(Optional) Don't emit messages or warnings if \code{TRUE}, default is \code{FALSE}}
}
\value{
a tibble with series, series group and last available data (if series has dates or \code{NA} if series does not have dates)
}
\description{
Get series to series group link and date of last available data point
}
