% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc.R
\name{get_boc_series_group_info}
\alias{get_boc_series_group_info}
\title{Get information on BOC time series groups}
\usage{
get_boc_series_group_info(series_group, refresh = FALSE)
}
\arguments{
\item{series_group}{A vector of series group identifiers}

\item{refresh}{(Optional) Refresh currently cached series if \code{TRUE}, default is \code{FALSE}}
}
\value{
a tibble with series metadata
}
\description{
The result is cached for the duration of the current R session
}
\examples{
\donttest{
series_group_info <- get_boc_series_group_info(c("FX_RATES_ANNUAL", "FSR-2017-JUNE-CHART-3"))
}
}
