% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc.R
\name{get_boc_series_info}
\alias{get_boc_series_info}
\title{Get information on BOC time series}
\usage{
get_boc_series_info(series, refresh = FALSE)
}
\arguments{
\item{series}{A vector of series identifiers}

\item{refresh}{(Optional) Refresh currenly cached series if \code{TRUE}, default is \code{FALSE}}
}
\value{
a tibble with series metadata
}
\description{
The result is cached for the duration of the current R session
}
\examples{
\donttest{
series_info <- get_boc_series_info(c("IEXE1001.CL", "IEXE4702"))
}
}
