% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boc.R
\name{list_boc_series_groups}
\alias{list_boc_series_groups}
\title{List available BOC time series groups}
\usage{
list_boc_series_groups(refresh = FALSE)
}
\arguments{
\item{refresh}{(Optional) Refresh currenly cached series if \code{TRUE}, default is \code{FALSE}}
}
\value{
a tibble with series information
}
\description{
The list of series is cached for the duration of the current R session
}
\examples{
\donttest{
series_groups <- list_boc_series_groups()
}
}
